test(ts#react-website#auth): refactor unit tests
